Extract DaxAction to its own file so it can be reused
