+++ android-design-system/design-system/src/main/java/com/duckduckgo/common/ui/compose/DaxAction.kt
@@ -1,0 +1,28 @@
+/*
+ * Copyright (c) 2026 DuckDuckGo
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.duckduckgo.common.ui.compose
+
+import androidx.compose.runtime.Immutable
+
+/** A simple button entry used by [com.duckduckgo.common.ui.compose.message.DaxMessageActions]
+ *  and [com.duckduckgo.common.ui.compose.snackbar.DaxSnackbar].
+ * */
+@Immutable
+data class DaxAction(
+    val text: String,
+    val onClick: () -> Unit,
+)
